003_LIVECell_Segmentation of Cells_U-Net/requirements.txt
altair==5.5.0
blinker==1.9.0
cachetools==6.1.0
click==8.2.1
gitpython==3.1.44
matplotlib==3.10.3
numpy==2.1.3
opencv-python==4.11.0.86
pandas==2.3.0
pillow==11.2.1
pyarrow==20.0.0
pydeck==0.9.1
requests==2.32.4
streamlit==1.46.1
tenacity==9.1.2
tensorflow-cpu==2.19.0     
toml==0.10.2
watchdog==6.0.0
